Extract findAllPackages
