+++ sk_sidekick/lib/src/commands/test_sidekick_context_command.dart
@@ -1,4 +1,5 @@
 import 'package:sidekick_core/sidekick_core.dart';
+import 'package:sk_sidekick/sk_sidekick.dart';
 
 class TestSidekickContextCommand extends Command {
   @override
@@ -9,10 +10,11 @@ class TestSidekickContextCommand extends Command {
 
   @override
   Future<void> run() async {
-    // print('repository: ${SidekickContext.repository}');
-    print(
-      'sidekickPackage: ${relative(SidekickContext.sidekickPackage.root.path)}',
-    );
-    print('entryPoint: ${relative(SidekickContext.entryPoint.file.path)}');
+    final sidekickPackage = SidekickContext.sidekickPackage;
+    print('sidekickPackage: ${relative(sidekickPackage.root.path)}');
+    final entryPoint = SidekickContext.entryPoint;
+    print('entryPoint: ${relative(entryPoint.file.path)}');
+    final repository = SidekickContext.repository;
+    print('repository: ${relative(repository.root.path)}');
   }
 }
